+++ CHANGELOG.md
@@ -20,6 +20,7 @@
    * ADDED: support for `filters` & `verbose` in `/tile` endpoint [#5806](https://github.com/valhalla/valhalla/pull/5806)
    * CHANGED: Timestamp log format changed from year/mo/dy hr:mn:sc.micros to year-mo-dy hr:mn:sc.nanosecnd [#5856](https://github.com/valhalla/valhalla/pull/5856)
    * CHANGED: Added `mjolnir.data_quality_dir` as optional config to control the directory for e.g. duplicateways.txt
+   * ADDED: `get_tile_ids_from_ring` to query by polygon in the Python bindings
 
 ## Release Date: 2026-01-15 Valhalla 3.6.2
 * **Removed**